Create voting context
