+++ lib/api_web/controllers/admin/admin_paper_vote_controller.ex
@@ -3,7 +3,8 @@ defmodule ApiWeb.Admin.PaperVoteController do
 
   alias Api.{Accounts, Accounts.User}
   alias Api.{Competitions.Category, Competitions.Team}
-  alias ApiWeb.{PaperVoteActions, ErrorController, PaperVote}
+  alias Api.{Voting, Voting.PaperVote}
+  alias ApiWeb.ErrorController
   alias Guardian.Plug.{EnsureAuthenticated, EnsurePermissions}
 
   action_fallback ErrorController
@@ -15,12 +16,12 @@ defmodule ApiWeb.Admin.PaperVoteController do
     category = Repo.get_by!(Category, name: c_name)
     user = Accounts.current_user(conn)
 
-    with {:ok, pv} <- PaperVoteActions.create(category, user),
+    with {:ok, pv} <- Voting.create_paper_vote(category, user),
       do: render(conn, "paper_vote.json", paper_vote: pv)
   end
 
   def show(conn, %{"id" => id}) do
-    render(conn, "paper_vote.json", paper_vote: PaperVoteActions.get(id))
+    render(conn, "paper_vote.json", paper_vote: Voting.get_paper_vote(id))
   end
 
   def redeem(conn, %{"id" => pv_id, "team_id" => t_id, "member_id" => m_id}) do
@@ -29,7 +30,7 @@ defmodule ApiWeb.Admin.PaperVoteController do
     member = Repo.get!(User, m_id)
     user = Accounts.current_user(conn)
 
-    with {:ok, pv} <- PaperVoteActions.redeem(pv, team, member, user),
+    with {:ok, pv} <- Voting.redeem_paper_vote(pv, team, member, user),
       do: render(conn, "paper_vote.json", paper_vote: pv)
   end
 
@@ -37,7 +38,7 @@ defmodule ApiWeb.Admin.PaperVoteController do
     pv = Repo.get!(PaperVote, pv_id)
     user = Accounts.current_user(conn)
 
-    with {:ok, pv} <- PaperVoteActions.annul(pv, user),
+    with {:ok, pv} <- Voting.annul_paper_vote(pv, user),
       do: render(conn, "paper_vote.json", paper_vote: pv)
   end
 end
