+++ lib/api_web/controllers/voting_controller.ex
@@ -3,7 +3,8 @@ defmodule ApiWeb.VotingController do
 
   alias Api.Accounts
   alias Api.Competitions
-  alias ApiWeb.{ErrorController, VotingView, VoteActions}
+  alias Api.Voting
+  alias ApiWeb.{ErrorController, VotingView}
 
   plug Guardian.Plug.EnsureAuthenticated, [handler: ErrorController]
     when action in [:upsert_votes, :get_votes]
@@ -13,7 +14,7 @@ defmodule ApiWeb.VotingController do
       :not_started ->
         ErrorController.call(conn, {:not_started, :not_found})
       _ ->
-        render(conn, VotingView, "info_begin.json", VoteActions.build_info_start())
+        render(conn, VotingView, "info_begin.json", Voting.build_info_start())
     end
   end
 
@@ -24,19 +25,19 @@ defmodule ApiWeb.VotingController do
       :started ->
         ErrorController.call(conn, {:not_ended, :not_found})
       :ended ->
-        render(conn, VotingView, "info_end.json", VoteActions.build_info_end())
+        render(conn, VotingView, "info_end.json", Voting.build_info_end())
     end
   end
 
   def upsert_votes(conn, %{"votes" => votes}) do
-    case VoteActions.upsert_votes(Accounts.current_user(conn), votes) do
+    case Voting.upsert_votes(Accounts.current_user(conn), votes) do
       {:ok, votes} -> render(conn, "upsert.json", votes: votes)
       {:error, error} -> ErrorController.call(conn, error)
     end
   end
 
   def get_votes(conn, _) do
-    case VoteActions.get_votes(Accounts.current_user(conn)) do
+    case Voting.get_votes(Accounts.current_user(conn)) do
       {:ok, votes} -> render(conn, "index.json", votes: votes)
       {:error, error} -> ErrorController.call(conn, error)
     end
