+++ test/controllers/admin/admin_paper_vote_controller_test.exs
@@ -2,7 +2,7 @@ defmodule ApiWeb.Admin.PaperVoteControllerTest do
   use ApiWeb.ConnCase
 
   alias Api.Competitions
-  alias ApiWeb.PaperVote
+  alias Api.Voting.PaperVote
   alias Guardian.{Permissions}
 
   setup %{conn: conn} do
